feat(signals): make signal emission pipeline source-agnostic
Extract the per-record processing loop (summarization, actionability,
emission with bounded concurrency, payload trimming) into a shared
pipeline.py and introduce a record_fetcher abstraction so each source
plugs in its own way of fetching records. Today only the data warehouse
fetcher exists; this prepares the ground for internal sources (Postgres,
ORM-backed, etc.) to plug into the same pipeline without forking it.
